Support for filling questions which are added "before" the most recently added question (#19)
In the D-Tree Postpartum form, there is a set of questions about danger signs. If the user answers "yes" to any of these questions, a new question is added asking about health facility follow-up. Standard stuff. But the new question actually appears "before" the danger sign question.

This change updates the form filler to continuously fill the "next unanswered question" instead of using an index to track the next unanswered question.

Also in this form on page 4, there were 3 "yes/no" questions but the form-filler erroneously found four questions on this page. Also fixed in this - will be published as medic-conf-test-harness@0.1.20
#19