Form Filler - Repeat sections with "+" button (#43)
Add support for "repeat sections" where the number of repeats is not specified in the form. The user experience is a "+" button that can be clicked to add repeat sections. For fill, pass an integer as form input and the button will be clicked that many times, then fill the new elements normally.