Reuse downloaded chromium when running npm ci
This is achieved by switching out puppeteer for puppeteer-chromium-resolver as a dependency.
Resolves #73.

puppeteer downloads chromium into the node_modules folder which is
cleared every time npm ci is invoked. This leads to always re-downloading
chromium (~110MB) which is a pain on slow or metered connections.
puppeteer-chromium-resolver on the other hand downloads chromium to a global
cache directory which persists previously downloaded artefacts across different
npm intstall/ci invokations.