v1.0 - harness.setNow() mocks global time via sinon.useFakeTimers (#12)
This is a proposed breaking change to the conf harness that changes "setNow" from mocking just "certain functions" to a global mocking of DateTime.

Currently, harness.setNow('2000-01-01') would mock the result returned by Utils.now() and mocks the time available within enketo forms, but it does not mock the global new DateTime(). This is particularly confusing when some code uses Utils.now(), but other does not (like contact summary which does not have that library available).

      const mockedTime = 1000;
      await harness.setNow(mockedTime);
      expect(new Date().getTime()).to.eq(mockedTime);
      await harness.clear();
      expect(new Date().getTime()).to.not.eq(mockedTime);
I believe the global mocking of time is a more consistent and intuitive experience, but this is a breaking change to the harness.

Published as medic-conf-test-harness@1.0.0-beta.0 for experimentation.