Link default contact information (#95)
The "default contact" is the contact that is "getting forms filled about them" be default. Currently users set the default contact through the harness.defaults.json file - via its content.contact (and to some degree user.parent). This BRAC test is linking the two copies together, so when you edit one you edit the other. And the test is right to do so - they intuitively should be the same thing. But they are each a unique deep copy.

This change deep copies the default only once - then shares that instance in the multiple places it is used. Therefore, editing the data in one place results in it being updated in both.