BREAKING - Data-structure and interface change to introduce "harness.subject" (#104)
harness.subject represents the contact that is being "acted on" or the "subject of the test". The fillForm() function simulates "completing an action" on the subject's profile page. The getTasks() function returns the tasks listed on the subject's profile page. The getContactSummary() function returns the contact summary displayed for the subject.

subject can be an object or a string. Assigning subject a string should be preferred as it will result in behavior closest to the CHT. The subject should be set to the _id of a contact document. The harness will combine and hydrate the appropriate documents.

# Breaking Changes
The harness.defaults.json file has been fully reworked
pushMockedDoc() now stores minified documents (that is the same the CHT stores)
Integrates with cht-core's lineage library for proper hydration and minification
getContactSummary() is now async to support hydration
harness.content.contact is removed by default