Run tests with --dev to set breakpoints and debug partner code (#103)
In the production environment, the command medic-conf compile-app-settings uses Terser to bundle partner code. This code is then wrapped in a proprietary syntax to make it a nools file. For projects using the declarative configuration system Nools plays no functional role - it is largely vestigial. However, it's proprietary syntax and parser prevent the use of SourceMaps.

This change replaces the RulesEmitter module with a mock. This mock provides similar (nearly identical) functionally but skips bundling and skips nools. The behaviour of the system should be nearly identical except some edge cases (eg. medic/cht-core#6506) but this provides some new experiences for test authors - like setting breakpoints and getting code coverage metrics for tests.