FillForm - Fails to answer questions with "falsey answers" like 0 (#128)
fillForm with answer of 0 Currently, the form-filler logic checks to see if the given answer is valued with if (!answer).... If this is true (meaning no value was given for the answer) then the logic will skip trying to set the answer on the form.

However, this logic will also be triggered if you try to set the 0 number literal as an answer (because !0 === true). The prevents 0 from being set as an answer (though, as a workaround, it is still possible to set '0'). This issue is illustrated in the added test case which will fail with the old logic check.